pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "clinical-trial-ae-triage"
version = "1.0.0"
description = "OpenEnv environment for pharmacovigilance adverse event triage and SUSAR detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "openai>=1.6.0",
    "python-dotenv>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages]
find = {include = ["env*", "server*"]}




